:memo: feat: double로 수정

See merge request s08-blockchain-contract-sub2/S08P22A209!131

+++ backend/src/main/java/com/donjo/backend/api/dto/support/SupportResponseDto.java
@@ -30,7 +30,7 @@ public class SupportResponseDto {
 
     private String replyMsg;
 
-    private Long amount;
+    private Double amount;
 
     private LocalDateTime arriveTimeStamp;
 
@@ -82,7 +82,7 @@ public class SupportResponseDto {
                 .supportType(support.getSupportType())
                 .fromMember(fromMemberAddress)
                 .toAddress(toMemberAddress)
-                .amount((long) (support.getAmount()/ Math.pow(10, 18)))
+                .amount((double) (support.getAmount()/ Math.pow(10, 18d)))
                 .arriveTimeStamp(support.getArriveTimeStamp())
                 .replyMsg(support.getReplyMsg())
                 .build();
@@ -93,7 +93,7 @@ public class SupportResponseDto {
                 .uid(support.getSupportUid())
                 .supportType(support.getSupportType())
                 .toAddress(toMemberAddress)
-                .amount((long) (support.getAmount()/ Math.pow(10, 18)))
+                .amount((double) (support.getAmount()/ Math.pow(10, 18d)))
                 .arriveTimeStamp(support.getArriveTimeStamp())
                 .replyMsg(support.getReplyMsg())
                 .build();
